clean up dataset page, and move half-baked to analysis
